requirements.txt
PySide6>=6.5
Pillow>=10.0
openpyxl>=3.1
numpy>=1.24
opencv-python-headless>=4.8
imagehash>=4.3
